+++ Paradigm/src/utopia/paradigm/shape/shape2d/area/polygon/Polygon.scala
@@ -37,6 +37,11 @@ case class Polygon(corners: Vector[Point]) extends Polygonic
 	override lazy val convexParts = super.convexParts
 	
 	
+	// IMPLEMENTED  -------------------
+	
+	override def center = super[Polygonic].center
+	
+	
 	// OTHER	------------------------
 	
 	/**