+++ Paradigm/src/utopia/paradigm/shape/shape2d/area/polygon/Triangle.scala
@@ -33,6 +33,8 @@ case class Triangle(origin: Point, side1: Vector2D, side2: Vector2D) extends Pol
 	
 	// IMPLEMENTED	----------------
 	
+	override def center = super[Polygonic].center
+	
 	override def transformedWith(transformation: Matrix3D) = mapCorners { _ * transformation }
 	override def transformedWith(transformation: Matrix2D) = mapCorners { _ * transformation }